backend/requirements.txt
astroid==2.4.1
click==7.1.2
colorama==0.4.3
ecdsa==0.15
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
future==0.18.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.9.7
pylint==2.5.2
python-jose-cryptodome==1.3.2
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
